separate methods for editVeterinariansPage
